+++ src/main/java/com/cMall/feedShop/config/SecurityConfig.java
@@ -2,10 +2,7 @@ package com.cMall.feedShop.config;
 
 import com.cMall.feedShop.user.infrastructure.security.JwtAuthenticationFilter;
 import com.cMall.feedShop.user.infrastructure.service.CustomUserDetailsService;
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -77,16 +74,5 @@ public class SecurityConfig {
         return source;
     }
 
-    @Bean
-    public OpenAPI customOpenAPI() {
-        final String securitySchemeName = "Bearer Authentication";
-        return new OpenAPI()
-                .addSecurityItem(new SecurityRequirement().addList(securitySchemeName))
-                .components(new Components()
-                        .addSecuritySchemes(securitySchemeName, new SecurityScheme()
-                                .name(securitySchemeName)
-                                .type(SecurityScheme.Type.HTTP)
-                                .scheme("bearer")
-                                .bearerFormat("JWT")));
-    }
+    
 }
